BUG#33640887 Load fails with: Invalid character in identifier
Loading a dump which contained a routine which was created when
ANSI_QUOTES sql_mode was in effect was failing with a cryptic message.

If a routine was created when ANSI_QUOTES was in effect, the dumped SQL
may contain identifiers quoted with double quotes. The code which is
executing the SQL was not expecting that, and reported invalid
identifier errors.

The code has been fixed, identifiers quoted with double quotes are now
correctly recognized.

An additional bug has been fixed in the dumper: if a table had a trigger
with an identifier which was required to be quoted, dumper did not quote
it and produced SQL statement which was invalid.

Change-Id: I23e787264f1602f9d208b14d37aba17381ddcf7e