WL#14769 Support for FIDO authentication in Shell
This patch adds support for FIDO authentication in the shell.

The authentication logic is provided by the MySQL client lib, however
a new command line option --fido-register-factor is introduced to enable
FIDO device registration in the Shell.

The authentication_fido_client plugin has an option to configure a print
callback to avoid the plugin printing data directly on stderr/stdout, to
configure this option a new event tracer plugin has been created which
is capable of detecting when authentication_fido is going to be used so
the plugin can be retrieved and the callback set.

Change-Id: Ieb573069ffff06dcb3c143b740129b02f9936a6d