BUG#33654278 mysqlsh crash while loading 2TB dump to instance
Shell could crash when loading the table data when two or more threads
attempted to execute HTTP requests at the time when the headers cache
was supposed to be cleaned. This issue affected only the load operations
from the OCI Object Storage done using the local OCI profile
(`osBucketName` option).

The cache cleaning operation is not thread-safe, the class responsible
for it is supposed to be created per each thread, however the
implementation did not take into consideration a scenario where object
created in one thread was passed to another thread. This lead to a
situation where the same instance was shared between multiple threads,
and when two threads performed an operation which triggered the cache
cleaning process, a crash occurred.

This issue has been fixed, the class responsible for cache cleaning is
no longer shared between threads.

Change-Id: Ic896cfc49782922ddcffad0fc4e21acb189f08a8