BUG#33517575 Dictionary used as the Python wrapper for a map should pass isinstance test
The implementation of the map wrapper used by Shell in Python mode has
been adjusted, it now inherits from the Python's dict (isinstance()
checks will return true). In order to avoid duplication of data and
ensure synchronization between Python's context and internal
representation of data, data is not stored in the base class, but in the
map wrapper. Due to this fact (and the fact that data is shared
between Python and JavaScript mode) it is only safe to store basic types
(i.e. numbers, strings, arrays, dictionaries) and native Shell objects
(i.e. Cluster, Session). As the base class is not used to store the
data, all methods of the list class have been reimplemented.

Limitations:
 - Order of insertion is not maintained, keys are always ordered
   alphabetically.
 - Only string type keys are supported, non-string keys are:
    - converted to a sting using str() when inserting multiple values
    - treated as missing when accessing an element
    - treated as an error (raising an exception) when setting a value

Change-Id: I501755fac7be9f42e72b062cd31a28a9bd37e717