Fixed failing AdminAPI tests.
Some tests that validate the output of a query to obtain the Group
Replication sysvars from performance_schema were failing due to a change
in the order on which the data was returned. The function to obtain the
data was assumming an order without enforcing it by making use of ORDER
BY. MySQL does not guarantee any order for SELECT statements.

This patch fixes that by introducing the usage of ORDER BY in that
statement and updating the tests validation files accordingly.

Change-Id: I5e45cfdc467cff78d1b051323a5870896db336a9