Fixing double registration of utils.debug help
The help for extension objects is expected to be registered only once, when the object
is registered, however this logic does not apply to the built in utils.debug object as
the object is re-created every time a new instance of the Mysql_shell class is created.

The scenarios where this could happen include:

- Tests that reinitialize the Mysql_shell object
- Plugin implementation that create threaded shell contexts

This patch prevents the double registration of the help for the built in utils.debug
object.

Change-Id: Ia083c7d4fcf182d6ca242ab62577b05174d9fe5d