BUG#33748812 group_replication_view_change_uuid on SR1 hotfix causing GR node failure
<Cluster>.rescan() verifies if group_replication_view_change_uuid is set
with a different value than the default one in the Cluster when the
target version is >= 8.0.27 and, if not, generates and set a value for
it on all members (using SET PERSIST). This action is automatic and
transparent to the users and the reason is that this is a requirement
for InnoDB ClusterSet.

However, Group Replication requires a full reboot of the Cluster in
order for the value to take effect. The AdminAPI warns the users about
it but won't perform the full reboot automatically.

Users that do not wish to use InnoDB ClusterSet or cannot affort to
perform a full reboot of the Cluster will have the auto-rejoin feature
of GR impacted in the sense that a member that auto-rejoins will fail to
join back the Cluster since its value of the sysvar is different than
the one currently in use in the Cluster (AUTOMATIC). This is
unnaceptable as it compromises HA.

For that reason, this patch changes the behavior of <Cluster>.rescan()
to not perform the change automatically. The command now warns the users
in such scenario and exposes a new setting 'updateViewChangeUuid' (false
by default) to allow users to explicitly indicate the change must be
performed.

Change-Id: I0b09126a085e469df1c3f5550a73f90ece7c8c7d