BUG#33699844 BUG#33697289 enforce consistency when dumping with consistent:true
BUG#33699844 Always enforce consistency when dumping if consistent:true
BUG#33697289 dumpInstance Errors out if the user doesn't have BACKUP_ADMIN privilege

The BACKUP_ADMIN privilege is no longer required to perform a consistent
dump of an 8.0+ server. Instead, if the BACKUP_ADMIN is not available
and LOCK INSTANCE FOR BACKUP cannot be executed (either because server
has a version lower than 8.0, or because user does not have this
privilege), an additional consistency check is performed during the
dump. If this check fails, a message is displayed and the dump is
aborted (in case of dumpInstance()) or continued (in the remaining
cases).

Additional changes:

The SHERR_DUMP_BACKUP_LOCK_MISSING_PRIVILEGES (52003) error code has
been changed to:
  SHERR_DUMP_CONSISTENCY_CHECK_FAILED 52003
  SHERR_DUMP_CONSISTENCY_CHECK_FAILED_MSG "Consistency check has failed."

Change-Id: I74519189c82664caf9fe4d2e79f41d15c9f92094