Bug #31964706 dba.configureInstance() lists log_bin twice on 5.7
Because "log_bin" has different properties in MySQL 5.7 and 8.0, when
gathering data to configure the instance, we weren't taking into account
that the var in question could be invalid on the server, but the cnf
file could already have it fixed. This would cause the same var to
appear twice, one for the server and the other for the cnf file.

This patch fixes the problem by not creating an entry for the server
(only the cnf file one is actually needed). Also, the patch changed how
we deal with vars that disable "log_bin" ("skip_log_bin" and
"disable_log_bin") by explicitly disabling them.

Other changes:
- removed magic number

Change-Id: I6f7ba6028d49ab398883fe38d2aff5135a3973aa