WL#14387 Support for S3 API file backend
New options
===========

The new options have been added to the following methods:

* `util.exportTable()`
* `util.importTable()`
* `util.dumpInstance()`
* `util.dumpSchemas()`
* `util.dumpTables()`
* `util.loadDump()`

`s3BucketName`
--------------

string (default: not set) - Use specified AWS S3 bucket for the location
of the dump.

If the `s3BucketName` option is used, the dump is stored in the
specified AWS S3 bucket, connection is established using the AWS
settings stored in the `.aws/credentials` and `.aws/config` files. The
directory structure is simulated within the object name.

`s3CredentialsFile`
-------------------

string (default: not set) - Use the specified AWS credentials file
instead of the one in the default location.

`s3ConfigFile`
--------------

string (default: not set) - Use the specified AWS config file instead of
the one in the default location.

`s3Profile`
-----------

string (default: not set) - Use the specified AWS profile instead of the
`default` one.

`s3EndpointOverride`
-----

string (default: not set) - Use the specified AWS S3 API endpoint instead
of the default one.

Handling of the AWS settings
============================

1. The following setting are read from the `config` file for the
   specified profile:

    * `aws_access_key_id`
    * `aws_secret_access_key`
    * `aws_session_token`
    * `region`

2. The following settings are read from the `credentials` file for the
   specified profile:

    * `aws_access_key_id`
    * `aws_secret_access_key`
    * `aws_session_token`

3. If there are credentials in both `credentials` and `config` files for
   the specified profile, the keys in the `credentials` file take precedence.
4. If either `aws_access_key_id` or `aws_secret_access_key` are missing,
   an exception is thrown.
5. If `aws_session_token` is missing, or is empty, it is not used to
   authenticate the user.
6. If `region` is missing, or is empty, a default value of `us-east-1`
   is used instead.

Changes in existing implementation
==================================
As a result of the refactoring required to add the AWS support, the
`progressFile` option of the `util.loadDump()` function always supports
both local and remote files. In order to use a remote file, the path must
use the HTTP(S) scheme, data is written to the remote file using HTTP PUT
requests.

Change-Id: I63a0cb87694403b3608098ec5f6876a9345f095e