Bug #32992693 Add instance check table without PK
When adding a new instance to a cluster, we check if there's any table
without a PK in the target instance, since these tables are not
replicated. However, when using clone as the recovery method, we can
ignore this check since all the data in the instance is going to be
wiped anyway.

This patch ignores the table PK check, when adding an instance to a
cluster and using the clone recovery method.

Change-Id: Ib8973c8191f9f7c9bce60845f5a6852ae89894d7