Bug #33538559 setPrimaryInstance doesn't support transactions timeout
The group_replication_set_as_primary UDF implementation was updated in
the server with a new option. This option provides a timeout to avoid
the UDF from blocking indeterminately if there are any ongoing
transactions. If this new timeout is reached, these ongoing transactions
are rolled back, allowing the UDF to proceed.

This patch adds a new option, "runningTransactionsTimeout", in
AdminAPI's command for primary switchover: Cluster.setPrimaryInstance().

Other changes:
 - updated a test in repl_account_management.js to prevent invalid reads
when executing queries in recently added instances.

Change-Id: Ia704bce234e07ab04a812f5b4adeccbb96f571c7