Bug #31775698 Shell claims primary change was successful
To change the primary member of a Cluster, the shell issues a call to a
specific UDF in the server. Due to the way UDFs work, the mechanism to
check for errors in a "normal" query doesn't apply to a UDF. This caused
the error to go unnoticed and the shell to report the call to set the
primary as successful.

This patch added a new error check specific for UDFs, which now catches
the error and shows it to the user.

Other changes:
- added support in unit tests to reference other server versions in a
generic way: MYSQLD_SECONDARY_SERVER_?
- cleanup a unique_ptr initialization

Change-Id: I569c0d5500c55e7578022a6c831af4385543d8de