BUG#33624751 Ensure that a FULLTEXT index is re-added first
Previously, when loading a dump with the `deferTableIndexes` option set
to `all`, the order of re-adding the indexes was not specified. With
this patch, FULLTEXT indexes are added first. Adding the first FULLTEXT
index to a table rebuilds the table if there is no user-defined
FTS_DOC_ID column. If other indexes are present, adding a FULLTEXT index
first is more optimal, as the other indexes do not need to be recreated.

Change-Id: I1644b63b9dcbdc77a409ce50241eabf2d2299371