Bug #29337169 Update expeltimeout check support
The default and maximum values of group_replication_member_expel_timeout
have been updated in versions 8.0.14, 8.0.21 and 8.0.20. Currently, the
AdminAPI only checks this option with the rules of version 8.0.13 (when
it was introduced).

This patch changes how the option is verified, by limiting the checks to
integers greater or equal to 0 (current or future max limits is done
server side), with the exception of versions <= 8.0.13, where we check
if the value is in the range [0, 3600]. The documentation for
"autoRejoinTries" was also updated to remove the mention of the range
check (which is also server side).

Change-Id: Id76dcf24d8e480cffd479e172a78a5a311aa0051