BUG#31964273 Prevent ReplicaSet from being added to a Cluster
When adding a new instance to a Cluster, an error is generated if that
target instance already belongs to an existing Cluster. However, if that
instance belongs to a ReplicaSet, no error is generated and the instance
is silently removed from the replica and added to the cluster.

This patch adds an error for this scenario, canceling the action if the
target instance belongs to a replica.

Other changes:
 - when creating a cluster, cluster set or replica, the shell now tries
to detect an unmanaged AR to better inform the user with a more
appropriate message.

Change-Id: Ib9a90f255a390e5b66fb4a53ab2d0a578e47c271