Fixed handling of leftover/zombie mysql instances in test runner:
Due to some changes in the MySQL server the shutdown sequence might
take longer and any attempt to remove the data directory while shutdown
is not complete results in high CPU usage and multiple writes to the
still open file descriptors that will increase disk space usage over
time (bug 33874936).

The test runner handling of sandbox leftovers attempts to connect to the
instance and if the connection is successfull it simply issues the query
"SHUTDOWN" and carries on. The next step is to delete the datadir and if
the instance is still shutting down the problem is seen.

This patch fixes and improves that by:

  - Verifying if the query is successfull
  - Waiting for the TCP port the server is listening on to be freed
  - Waiting for the server process to be ended, by waiting for the pid
  file to be deleted which means the server shutdown is complete

Change-Id: Ie14f8106f6a1e89d04d1f5bf8bed0aabf4e23bd7