BUG#33907474 Unable to import OCI in linux platforms with bundled python
This patch removes the generation of the easy-install.pth.

The easy-install.pth was required when the python dependencies were being bundled
as .egg files to make the addition of them into the PYTHONPATH automatic.

Now that they are being installed from .whl packages this file is causing the paths
to each bundled package to be unnecessarily added to the PYTHONPATH.

At the mimetypes module, the following import is expected to fail with a "module not
found error" in linux (which is properly handled):

from _winapi import _mimetypes_read_windows_registry

However, the paramiko package includes a package named the same and it is found,
causing the import to fail with the error described on the bug report, and so
causing the import failures.

Change-Id: Ic76a666c80ccd666496f7307a18a2d49990dc8f0