BUG#33602309 Dba.upgradeMetadata: invalid stoi argument (ArgumentError)
Several Router management related operations such as .listRouters() or
even a Metadata upgrade would fail if Router was bootstrapped to use
sockets and/or to disable TCP ports.

The problem happens at Router metadata unserialization on which the
AdminAPI assumes that  "ROEndpoint", "RWEndpoint", "ROXEndpoint",
and "RWXEndpoint" are strings representing numeric values and uses
std::stoi() to convert them to integers.

This patch fixes that by not converting those values to integers and
simply treat them as strings.

Added regressions tests.

Change-Id: Ib1d13f78443dfa1de38fcf956f8bf235ccc6fa60