Bug #33235502 Unused cluster recovery accounts left behind
In cases where the user doesn't wait for the clone recovery method to
finish when an instance is added to a cluster, two scenarios can happen:
 - it's possible that the recovery account, when the instance is later
removed from the cluster, to remain behind (and unused)
 - instances still belonging to the cluster aren't actually using the
correct recovery account that was supposedly created for each one

The patch solves this by adding warning messages to the instances in
Cluster.status(), in the "instanceErrors" field and also corrects these
issues in a call to Cluster.rescan().

Another unrelated issue in Cluster.rescan(), where fixing non-existing
recovery accounts wasn't taking into account the cluster allowed host,
was also fixed.

Change-Id: I0bf768067eb7e2f3878cd7cfe1d1c4199845da69