BUG#33850528 rebootClusterFromCompleteOutage() not using view_change_uuid set by rescan()
group_replication_view_change_uuid is a required setting when using
InnoDB ClusterSet that is verified by Cluster.rescan(). Users are given
the option to let the AdminAPI generate and set a value for it in
Cluster.rescan() which will then be used by the Cluster whenever it is
restarted and rebooted from complete outage with
dba.rebootClusterFromCompleteOutage().

However, .rebootClusterFromCompleteOutage() does not inspect nor changes
the value in place for _view_change_uuid. Considering that persisted
settings are only effective after a server restart, if users don't
restart the servers of the Cluster but simply stop Group Replication on
every member of if, which is in fact a Cluster shutdown, the Cluster is
rebooted using the default value for _view_change_uuid (AUTOMATIC).
This leds to an endless loop of .rescan() detecting the missing setting
and .rebootClusterFromCompleteOutage() not using the generated value.

This patch fixes that by making .rebootClusterFromCompleteOutage()
check if _view_change_uuid was persisted by querying
performance_schema.persisted_variables, to then use it when rebooting
the cluster. Instances rejoined to it during the process or even later,
will use the value being used by the seed member.

Change-Id: I30cac03a8d507b8463f601904c9b4ad414ea71a8