Bug#33016337 CAN'T SET PRIMARY NODE AS MISMATCH BETWEEN MEMBER_HOST AND INNODB METADATA
setPrimaryInstance() (and maybe other functions) will lookup the
address/endpoint of the target instance and use the address stored in
the metadata for the intended operation.

If the user changes the value of report_host after the instance is
already in the cluster, the metadata will become out-of-sync causing
problems in such operations.

Cluster.rescan() was updated to detect and repair such cases
automatically. instance_name, address and all endpoint addresses are
updated in the metadata from values queried from the instance.

Cluster.status() was updated to detect and report the issue in
instanceErrors.

Change-Id: I3680a6caf66014b1f3f723e5343ef65b3f22845d