BUG#33630808 Simplify or automate the GR recovery user management
When an instance is added to a cluster using clone as the recovery
method but the user decides to not wait for the completion of the
recovery, by either canceling the progress using ctrl-c or using the
option `waitRecovery: 0`, the recovery account created when adding that
instance will become a leftover in the cluster whenever the instance is
removed with cluster.removeInstance().

This happens because of the internal handling of clone recovery:

  - After clone is complete, a CHANGE MASTER is done to ensure the right
recovery account is used by the target instance (otherwise the account
from the seed is used since clone copies everything including the
replication configurations)

  - Every newly created recovery account is stored in the Metadata
referring to the instance being added to ensure that whenever an
instance is removed from the Cluster, if there are 1 or more instances
still using that instance's recovery account, the account is not dropped

When the user decides to not wait for the completion of the recovery,
the CHANGE MASTER command is not done so the instance will use its donor
account. Whenever the instance is removed, the AdminAPI sees that
there's another instance still using that account (the donor) so won't
drop it leaving a leftover account.

The reason why the account becomes a leftover is that the logic to drop
the account didn't consider that the recovery account in use by the
instance being removed might not be the same account created for the
instance when it joined the cluster.

This patch extendes the account removal logic to consider that scenario.

Added regression tests.

Change-Id: I3fa486c35d93cdf305c3c2d9b845b12af398e548