BUG#33914146 race condition in dba.upgradeMetadata()
When an outdated Router is upgraded as part of an upgrade scenario, and
the Metadata schema upgrade process starts, the first step is to upgrade
the router accounts to allow router perform updates in the new metadata
version. Only then, Router can update the Metadata with the missing
version information to allow the upgrade process to complete
(BUG#33825480).

However, Router is only able to do such update after the privileges were
given and a new metadata cache reload happens (TTL). For that reason,
users face a race condition in the dba.upgradeMetadata() command that
would report Routers as being outdated since the Metadata wasn't updated
yet. Performing 1 or more retries is the workaround, however, this is
misleading to the user and can be avoided.

This patch fixes that by adding a loop with a timeout of 4 seconds for
the retrieval of Router's metadata information before returning the
results to the user. Considering that Router's default TTL is 0.5
seconds, this should overcome the race-condition in most cases. On top
of that, added a new NOTE message indicating this possibility and
suggesting a retry/restart if Router's TTL is high.

Change-Id: I035a6d2d4890c5b34e80eedbd4951986a2e8ac71