BUG#33893435 cluster.setPrimaryInstance fails
As per RFC 1034, domain names should be compared in a case-insensitive
manner, or in other words, "hostname" is the same as "HOSTNAME".
Currently, hostnames / addresses aren't compared in this way, which
means that operations where the hostnames differ (in terms of
case sensitivity) will fail.

This patch fixes hostname comparisons so that the rules specified in the
RFC are correctly followed. This means that, hostnames that aren't case
identical, now compare correctly as being the same.

Change-Id: Ic1b10101b43e16100688049af2c1151396f521ef