BUG#33949419 Prompts in JSON mode missing new line
With the enhancements of the Shell prompts, the produced JSON documents stopped
including a new line at the end in a way that the input as read from the user is
done on the next line, this is required for the proper integration into external
systems that rely on getting complete JSON documents per line.

This is a regression bug.

Change-Id: If9320fe36c2088447e2268c3ae50124169e17c4d