Post fix for BUG30548447
A regression was introduced in the patch that solved BUG#30548447.

The regression happens when, in interactive mode, the user is prompted
to setup a cluster based on an existing replication group, answers
"yes", and the shell returns an error saying that the "adoptFromGR"
option must be set to true. This essentially ignores the response.

This particular use case went unnoticed because, in the original patch,
a code cleanup was also made (where tests where upgraded to the newest
version), and the cleanup precisely missed the tests that covered this
scenario.

This patch fixes the regression and adds the tests.

Change-Id: Id513cbd15947c77cbd212995a4329dc3ce9b06f0