BUG#33976259 Do not use BETWEEN when upper and lower bounds are the same
When dumping tables with compound primary keys, chunker could produce
chunking conditions in form of: `column` BETWEEN X AND X (the same lower
and upper bounds). Such conditions combined with the ORDER BY clause
appear to cause a filesort, which means that when multiple chunks of a
big table were dumped at once, OS would run of disk space and dump would
fail with:
  MySQL Error 9 (HY000): Unexpected EOF found when reading file
  '/tmp/MYEqko1C' (Errcode: 175 - File to short; Expected more data in
  file)

Chunker code has been adjusted, if the lower and upper bounds are the
same, a `column` = X condition is used instead.

Change-Id: I4212da6d8d300442d948a04acf54c0a458235c79