BUG#33972939 Unexpected BE error for query: month must be in 1..12
Date/Time columns in the database are converted into python datetime/date times
when the data is read using the python API.

The conversion did not consider the fact that MySQL can accept invalid datetime/date
values by relaxing the SQL_MODE, in such scenario, the data type conversion is raising
the python error about invalid values for datetime/date.

This patch addresses this problem to ONLY do the conversion when possible, in case
of errors, the conversion will take the string representation of the original value.

Change-Id: I1a4a3b24f177f4b8df1b28126dbb146b0fd3c84c