WL#14765 BUG#34053105 BUG#34053126 BUG#34053084 BUG#34058442 BUG#34068796
AdminAPI - Support for new GR protocol stack

In MySQL 8.0.27, Group Replication introduces the ability to use the
existing Server facilities to establish connections among Group members
instead of using the internal XCom communication infrastructure. It is
called the 'MySQL' communication stack.
This allows a simpler and safer setup as it obsoletes the usage of IP
allowlists and removes the explicit need of having another network
address used for internal GCS communication, and introduces user-based
authentication.

This patch adds support in the AdminAPI for the new communication stack by:

  - Adding support to choose which GR communication stack must be used
  when creating a Cluster or a Replica Cluster, by exposing a new option
  named 'communicationStack' supporting the 2 values: 'XCOM' or 'MYSQL'

  - Ensures 'MySQL' is the default communication stack in use by
  Clusters running a version >= 8.0.27.

  - Automatically and transparently generates and sets all
  configurations required for both communication stacks at cluster
  creation time, and topology changes (add/rejoin), without any user
  intervention.

  - Validates the usage of the communication stack when required and
  ensures the right settings are in place in the commands that perform
  Cluster topology changes.

  - Lists the communication stack in use by a Cluster in
  .status({extended: 1) and .options()

  - Allows migrating an existing Cluster to a different communication
  stack when rebooting it from complete outage by exposing a new option
  named 'switchCommunicationStack'.

  - Allows setting a value for group_replication_local_address in
  Cluster.rejoinInstance() by extending the command with a new option
  'localAddress'.

It also introduces a new requirement for InnoDB Cluster and ClusterSet:

  - If 'group_replication_tls_source' is used, it must be set to
  'mysql_admin'.

Finally, it adds a new check in Cluster.rescan() to verify whether there
is any Cluster member using a recovery account that is not the one
created for itself and ensure the right account is used. This ensures
instances that fall in that category (added with clone recovery and
waitRecovery:0) are able to auto-rejoin the Cluster.

And, extends Cluster.rejoinInstance() and
dba.rebootClusterFromCompleteOutage() to also perform the instance
checks for Cluster compliance likewise Cluster.addInstance() does.
Even though it's assumed that an instance rejoining or rebooting a
cluster will have its configurations in accordance since the instance
was part of a Cluster before, there's nothing to prevent users to change
configurations before attempting those checks so the AdminAPI must
validate the configurations again to avoid further errors and problems.

Change-Id: I80bb657238ec99558a44ff82fef4c447ec270b2f