BUG#34058286 Log file permissions warning doesn't honor JSON output format
Some non critical warnings were introduced regarding the validation of the
log file permissions, these validations occur before the shell output handling
is initialized, it was simply impossible to honor the formatting output
configured in the shell.

Since these are not critical checks, it is possible to delay the printing of the
warnings for when the output handler is initialized, honoring the output
format configured on the shell call.

Change-Id: Ife3e67edd9727b489f0da2f15212e7f86723221d