BUG#34013718 VIEW CHANGE UUID CAUSES FAILURE TO SECONDARY CLUSTER DURING CLUSTERSET SWITCHOVER
Attempting a switchover of the Primary Cluster in a ClusterSet, on which
the target Cluster had a diverging GTID-SET due to View Change Events
(VCLEs) and all binary logs were purged, breaks the ClusterSet
replication channel and the Replica Cluster drifts apart.

The root cause of the issue is the switchover not performing the
reconciliation of the GTID-SET in regards to VCLEs and, since all binary
logs were purged, the missing GTIDs cannot be replicated to the Replica
Cluster. The lower-level cause of the missing reconciliation is a wrong
comparison of the GTID-SETS of the Primary and Replica Clusters that was
filtering VCLEs at that stage when it shouldn't.

<ClusterSet>.rejoinCluster() is also affected.

This patch fixes that by ensuring the reconciliation is performed
accordingly when needed.

Added regression tests.

Change-Id: I7bf6076a94e75a48cedb0741bb2fcf716e28cb30