Fixed possible undefined behavior:
The 2nd arg of Group_replication_options::read_option_values() is a bool
which is being passed as a deref of an std:optional(). That variable is
not initialized in some cases, resulting in undefined behavior.

This can affect <Cluster>.rejoinInstance() and
dba.rebootClusterFromCompleteOutage() when the communication stack
feature is being used.

Change-Id: I23a03d07d32156693eaa8680cf4899f925885082