BUG#34122867 description of lastCheckIn in .listRouters() helper is incorrect
In 8.0.30, with BUG#34092926, the whole purpose of Routers' lastCheckIn
changed. Router does not update the Metadata periodically with the
lastCheckIn timestamp anymore, it will do it only once at startup time.

For that reason, this patch updates the helper documentation of
`lastCheckIn` to warn the users about that behaviour change. The update
covers all commands that use it:

  - <Cluster>.listRouters()
  - <ReplicaSet>.listRouters()
  - <ClusterSet>.listRouters()

Change-Id: I71891b4dd45fd474f9b1f464c94b79a744d64687