BUG#33922830 Using upgradeCommProtocol in rescan() does not upgrade to the latest version
In 8.0.27, a new version of the Group Replication communication protocol
was introduced (8.0.27), however, the AdminAPI wasn't extended to
identify and allow upgrading the protocol to the latest version
available.

For that reason, using cluster.rescan({upgradeCommProtocol: true}) when
an upgrade to version 8.0.27 was possible was a no-op. Also, informative
messages in removeInstance() whenever the protocol is detected to be
upgradable ignored the fact that an upgrade was possible.

This patch fixes that by extending the AdminAPI to consider the new
version of GR's communication protocol.

Added regression tests.

Change-Id: If8f572a1a42fa23b8d0b618394fa361a78780fab