BUG#34110375 Not possible to change sql_generate_invisible_primary_key in MDS
Support for automatically creating invisible PKs was added in MySQL
8.0.30, this behaviour is controlled using the
sql_generate_invisible_primary_key system variable. Shell was setting
this session variable while loading a dump to an appropriate value
(based on the compatibility options used during the dump and the value
of the createInvisiblePKs option) regardless of the current value of
this variable. This led to load failures in MDS, as this variable cannot
be changed by the customers (they lack the SESSION_VARIABLES_ADMIN
privilege).

The code has been changed to only set sql_generate_invisible_primary_key
variable if its global value is different from the one requested by the
user (either implicitly, using the compatibility options, or explicitly,
using the loadDump() option). It is still possible that this variable is
going to be set when loading a dump into MDS, which will fail due to the
reasons explained above. This would mean that i.e. "ignore_missing_pks"
compatibility option was used during the dump, and user is trying to
load it into an HA enabled MDS instance.

Change-Id: I70f84ec34c3fc3f1b8bd8bedd900c02439fcc895