BUG#33951851 Unable to load a dump, shell behaves as if metadata files were missing
Shell was unable to load a dump using linux-glibc2.12-x86-64bit binary
if dump was stored on a filesystem which does not support d_type field
of the dirent structure (i.e. xfs with ftype parameter set to 0). Bug
was caused by GCC older than 8.4.0, GCC in these versions does not
handle such filesystems, and it's not possible to detect the file type.
The newer versions fall back to stat().

When shell lists the directory contests, it's only interested in the
regular files, ignoring symlinks, sockets, etc. If file type cannot be
determined, it is ignored as well. The list used by the loader is
incomplete, metadata files are not read and parsed, and nothing is
loaded.

As a fix, if shell is compiled by affected GCC, it falls back to stat()
as well.

Fix was verified using an extracted listing code, as follows:

  dd if=/dev/zero of=virtual.img bs=1M count=100
  mkfs.xfs -f -n ftype=0 -m crc=0 virtual.img
  sudo mkdir /mnt/virtual
  sudo mount -t auto -o loop /opt/virtual.img /mnt/virtual

  $ ll /mnt/virtual/dir/
  total 0
  drwxrwxr-x. 2 user user 42 May  6 12:24 ./
  drwxr-xr-x. 3 user user 16 May  6 12:24 ../
  -rw-rw-r--. 1 user user  0 May  6 12:24 1.txt
  -rw-rw-r--. 1 user user  0 May  6 12:24 2.txt
  -rw-rw-r--. 1 user user  0 May  6 12:24 3.txt

  $ ./list_files.8.3.1 /mnt/virtual/dir/
  0 files in '/mnt/virtual/dir/':

  $ ./list_files.8.3.1.fixed /mnt/virtual/dir/
  3 files in '/mnt/virtual/dir/':
    3.txt
    1.txt
    2.txt

Change-Id: I719d21cfc859ee6df0c529f2aaae5af58923e0d3