Bug #33707127 MySQL shell preupgrade check failed due to MySQL schema nul, prn, and aux
The upgrade checker is reporting false positives when reserved names are used for schemas
and tables. When a reserved name is used, the name will be registered in I_S.INNODB_SYS_TABLES
by appending @@@ to the name, but in I_S.TABLES the original name is used. This handling causes
the upgrade checker to report the table as false positive for schema consistency.

This patch removes the @@@ from schema/table names at I_S.INNODB_SYS_TABLES so the names
match with those in I_S.TABLES.

Change-Id: Ifcf1a878a0d413bc318ff5dabe4628f8e5a1e7db