Fixing seg-fault related crash (2/3)
This patch includes several small fixes that try to correct the cause of
a random seg-fault (which seems to affect several architectures).

The main change is the cleanup of Python objects by encapsulating them
into py::Release and/or py::Store wrappers to ensure that they are not
misused.

Other changes include:
 - protected Value assignment operators from self-referencing
 - added noexcept to move ctor and assignment to avoid copies

Change-Id: I632a318c17e0d489a21dc32bfaf59213e32ce9ae