BUG#34069295 removeCluster() does not reconcile GTID-SET for VCLEs
When a Replica Cluster is created on an instance that was the only
member of a removed Replica Cluster, the instance will contain errant
transactions that did not originate from the Primary Cluster and cause
the ClusterSet replication channel to break.

Those transactions are the View-Change-Events that were created as part
of the bootstrap of the Cluster to which the instance belonged.

If the Replica Cluster is created on an instance that was a secondary
member of the removed Cluster, the issue is not seen.

This bug was caused by a missing reconciliation of VCLEs in the primary
member of the removed Cluster. When the Cluster had secondary members
the reconciliation was done after the removal of each secondary member,
however, if the Cluster only had 1 member (the Primary) no
reconciliation was done.

This patch ensures reconciliation is performed always and before
completing the dissolve of the Cluster, that is, before removing the
Primary member.

Added regression tests.

Change-Id: Ic843db4113516fc81474d85e151fa8440a719be3