BUG#34097411 Shell should provide default credential manager if not available in the system - postfix
If the mysql_config_editor was not bundled with the Shell package, the
login-path helper would stop working (it was unable to find the system
mysql_config_editor even if it was present). This has been fixed.

DEB Builds: the handling of the bundled mysql_config_editor was being
ignored by the cmake  configure step because it was defined too late,
relocating the handling fixes the DEB builds.

RPM Builds: in most platforms the INSTALL_LIBEXEC dir is named "libexec"
but in some cases it is still "lib" (i.e. OpenSuse), this is determined
by the RPM packaging available on the system. This patch allows the RPM
build configure step to define the INSTALL_LIBEXEC directory to be used
in the packages.

Change-Id: Iba6f32327c16692508b49276b2b2c6f29b9adfba