BUG#33976718 loadDump to utilize parallel DDL for indexes
MySQL server 8.0.27 introduced support for parallel index creation, this
patch adds code to better utilize this feature.

Previously, loader would add indexes to a table one by one, with this
change all indexes in a table are added at once. The only exception are
SPATIAL indexes, which are added in a separate batch, but only if table
doesn't have virtual columns or FULLTEXT indexes.

Since adding all indexes at once may result in
ER_TEMP_FILE_WRITE_FAILURE errors (failure to write to directory
specified by innodb_tmpdir server variable, i.e. due to a full disk),
an additional handling of this error has been added: query is retried
with a smaller number of indexes.

As an additional change, the scheduling algorithm takes into account the
values of innodb_parallel_read_threads and innodb_ddl_threads server
variables, and the impact that a single ALTER TABLE ... ADD INDEX
statement would have on the target instance. Based on the computed
weight, ALTER TABLE tasks are executed only when an appropriate number
of worker threads are idle. This allows to balance the load of the
target instance.

Note: both previous and current implementation allow for multiple
index creation tasks to be executed at the same time, the main
difference is that previously a task would add a single index per
statement in a loop for all indexes to be added, while currently a task
attempts to add all indexes at once in a single statement.

Change-Id: I14d185863ded35ea983be2b04d3c0a75f8cda226